fix(QrCode): export per file
