+++ components/index.js
@@ -47,6 +47,7 @@ export { default as convertImgUrlToDataURLviaCanvas } from './utils/convertImgUr
 export { default as dateDiff } from './utils/date/diff'
 export { default as dateFns } from './utils/dateFns/dateFns'
 export { default as debounce } from './utils/debounce/debounce'
+export { default as downloadQR } from './utils/downloadQR/downloadQR'
 export { default as decodeToken } from './utils/auth/decodeToken'
 export { default as formatDate } from './utils/date/formatDate'
 export { default as formatDateStandard } from './utils/date/formatDateStandard'
@@ -188,7 +189,7 @@ export { default as Messaging } from './organisms/messaging/container/container'
 export { default as MessagingActions } from './organisms/messaging/communication/messagingActions'
 export { default as PasswordChange } from './organisms/passwordChange/passwordChange'
 export { default as Pricing } from './organisms/pricing/pricing'
-export { Qrcode, downloadQR } from './organisms/qrCode/qrcode'
+export { default as Qrcode } from './organisms/qrCode/qrcode'
 export { default as Register } from './organisms/register/register'
 export { default as Schedule } from './organisms/charts/scheduleChart/scheduleChart'
